% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildLogLikelihood.R
\name{BuildLogLikelihood}
\alias{BuildLogLikelihood}
\title{Builds a log-likelihood function for a given pair of pdf and data.}
\usage{
BuildLogLikelihood(pdf, data)
}
\arguments{
\item{pdf}{A probability density function that expects a pair (x, theta) and perhaps a log.p value.}

\item{data}{A vector of data.}

\item{theta}{A vector of parameters for pdf.}
}
\value{
A log-likelihood function expecting a parameter vector.
}
\description{
Builds a log-likelihood function for a given pair of pdf and data.
}

