% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotg.R
\name{dcotg}
\alias{dcotg}
\title{Density function for the Co-Tangent Gumbel probability distribution.}
\usage{
dcotg(par, x)
}
\arguments{
\item{par}{A vector of parameters. Only the first two will be used as alpha and beta.}

\item{x}{Point where to evaluate the density.}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Density function for the Co-Tangent Gumbel probability distribution.
}
\examples{
pdf_cotG(c(1, 1), 0)
pdf_cotG(c(0.1, 0.8), 0.5)
pdf_cotG(c(1.5, 1), 1)
pdf_cotG(c(1, 0.5), 1.5)


}
