% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustOptim.R
\name{RobustSANN}
\alias{RobustSANN}
\title{Does lots of crap.}
\usage{
RobustSANN(likelihood, starts, data, method = "PSO", domain = c(0, Inf),
  mle = NULL, ...)
}
\arguments{
\item{likelihood}{.}

\item{starts}{.}

\item{data}{.}

\item{method}{.}

\item{domain}{.}
}
\value{
Lots of crap.
}
\description{
Does lots of crap.
}
